MakeBooking MultipleServices tests for greedy allocation
(cherry picked from commit 37e315da1c8ca65ad753864b7a70d440a3653a9a)
