+++ src/api/Nhs.Appointments.Core/AvailabilityCalculator.cs
@@ -24,9 +24,7 @@ public class AvailabilityCalculator(
         var filteredSessions = service == "*"
             ? allSessions
             : allSessions.Where(s => s.Services.Contains(service));
-
-        //TODO bookings not filtered by service in SingleService code path.
-        //what happens if we enable MultiServices then disable?
+        
         var bookings = await bookingDocumentStore.GetInDateRangeAsync(from.ToDateTime(new TimeOnly(0, 0)),
             until.ToDateTime(new TimeOnly(23, 59)), site);
